+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First letter golden gradient.
 - Standalone page option for displaying collections.
+- Search mechanism.
 ### Changed
 - Change the demographics' class and subclass to customizable.